+++ snowpack.config.js
@@ -1,8 +1,0 @@
-module.exports = {
-  optimize: {
-    entrypoints: ['./src/index.js'],
-    bundle: true,
-    minify: true,
-    target: 'es2015',
-  },
-};