+++ src/index.js
@@ -1,5 +1,1 @@
 export { createObservable } from './createObservable';
-
-export default {
-  createObservable,
-};