+++ CHANGELOG.md
@@ -7,8 +7,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2021-02-23
+
+### Change
+
+- Swapped [snowpack](https://github.com/snowpackjs/snowpack) bundler for [rollup](https://github.com/rollup/rollup) by [@lukasduspiva](https://github.com/lukasduspiva).
+
+### Fixed
+
+- Fixed `README.md`, `package.json` and `CHANGELOG.md` by [@lukasduspiva](https://github.com/lukasduspiva).
+
+## [0.1.0] - 2021-02-23
+
 ### Added
 
 - Kicked off the package with `createObservable` function by [@lukasduspiva](https://github.com/lukasduspiva).
 
-[unreleased]: https://github.com/lukasduspiva/simple-observables/compare/v1.0.0...HEAD
+[unreleased]: https://github.com/lukasduspiva/simple-observables/compare/v0.2.0...HEAD
+[0.1.1]: https://github.com/lukasduspiva/simple-observables/compare/v0.1.0...v0.2.0
+[0.1.0]: https://github.com/lukasduspiva/simple-observables/releases/tag/v0.1.0